refactor: move Postal::BounceMessage to app/models/bounce_message
